+++ go/cs/reservation/segment/admission/stateful/stateful.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package stateless
+package stateful
 
 import (
 	"context"
